store/requirements.txt

Flask==2.2.3
Werkzeug==2.2.3
Flask-PyMongo==2.3.0
pymongo==4.3.3
PyJWT==2.6.0
python-dotenv==1.0.0